src/pages/sign-in-page/sign-in-page.test.tsx
import { render, screen } from '@testing-library/react';
import userEvent from '@testing-library/user-event';
import { withHistory, withStore } from '../../utils/mock-component';
import SignInPage from './sign-in-page';
import { makeFakeStore } from '../../utils/mocks';
describe('Component: SignInPage', () => {
  it('should render correctly', () => {
    const loginText = 'Email address';
    const passwordText = 'Password';
    const store = makeFakeStore();
    const { withStoreComponent } = withStore(<SignInPage />, store);
    const preparedComponent = withHistory(withStoreComponent);

    render(preparedComponent);

    expect(screen.getByText(loginText)).toBeInTheDocument();
    expect(screen.getByText(passwordText)).toBeInTheDocument();
    expect(screen.getByRole('button')).toBeInTheDocument();
  });

  it('should render correctly when user enter login and password', async () => {
    const loginElementTestId = 'loginElement';
    const passwordElementTestId = 'passwordElement';
    const expectedLoginValue = 'may@gmail.com';
    const expectedPasswordValue = '123456';
    const store = makeFakeStore();
    const { withStoreComponent } = withStore(<SignInPage />, store);
    const preparedComponent = withHistory(withStoreComponent);

    render(preparedComponent);
    await userEvent.type(
      screen.getByTestId(loginElementTestId),
      expectedLoginValue
    );
    await userEvent.type(
      screen.getByTestId(passwordElementTestId),
      expectedPasswordValue
    );

    expect(screen.getByDisplayValue(expectedLoginValue)).toBeInTheDocument();
    expect(screen.getByDisplayValue(expectedPasswordValue)).toBeInTheDocument();
  });
});
